// src/app/interfaces/okm_signovital.interface.ts
export interface OKM_Signovital {
  name?: string;
  parent?: string;
  parentfield?: string;
  parenttype?: string;
  ate_id?: string;
  pac_historiaclinica?: string;
  prosal_cedopasaporte?: string;
  sigvit_fecha?: string;
  sigvit_altura?: number;
  sigvit_peso?: number;
  sigvit_imc?: number;
  sigvit_perimetroabd?: number;
  sigvit_perimetrocef?: number;
  sigvit_abdomen?: number;
  sigvit_presionasis?: number;
  sigvit_presionadia?: number;
  sigvit_presionmedia?: number;
  sigvit_temperatura?: number;
  sigvit_frecardiaca?: number;
  sigvit_satoxigeno?: number;
  sigvit_frerespiratoria?: number;
  sigvit_glucosacap?: number;
  sigvit_hemoglobina?: string;
  sigvit_hemoglobinacorre?: string;
  sigvit_colorlengua?: string;
  sigvit_reflejo?: string;
  sigvit_pulso?: string;
  sigvit_fotoact?: string;
  sigvit_observacion?: string;
  sigvit_cita?: string;
  unimed_codigo?: string;
  sigvit_tallaedad?: number;
  sigvit_tallaedadestado?: string;
  sigvit_pesoedad?: number;
  sigvit_pesoedadestado?: string;
  sigvit_pesotalla?: number;
  sigvit_pesotallaestado?: string;
  sigvit_imcedad?: number;
  sigvit_imcedadestado?: string;
  sigvit_sololechematerna?: string;
  sigvit_solidosemisuave?: string;
  sigvit_hierromineral?: string;
  sigvit_vitaminaa?: string;
  sigvit_xqvitaminaa?: string;
  sigvit_xqhierrominerales?: string;
}